requirements.txt
openrgb-python==0.2.10
